requirements.txt
Flask==3.1.0
firebase-admin==6.3.0
